+++ CHANGES.rst
@@ -6,7 +6,7 @@ Change log for risclog.zestreleaser.notify
 1.0.4 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Add hookshot integration.
 
 
 1.0.3 (2023-07-05)